+++ calcipy/dev/noxfile.py
@@ -279,6 +279,7 @@ if _HAS_TEST_IMPORTS:  # pragma: no cover  # noqa: C901
         session.install('wily', '--upgrade')
         logger.warning('FYI: Can only be run when all changes are checked in or stashed')
         # All possible metrics can be listed with: wily list-metrics
+        # TODO: How are the additional metrics configured?
         # operators = ','.join([
         #     'raw.loc', 'raw.lloc', 'raw.sloc', 'maintainability.rank', 'maintainability.mi', 'cyclomatic.complexity',
         #     'halstead.h1', 'halstead.vocabulary', 'halstead.length', 'halstead.volume', 'halstead.difficulty', 'halstead.effort',