+++ docs/docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Feat
 
+- replace toml with tomli (#74)
 - move wily to nox to reduce version conflicts
 - add pip-audit
 - add attrs_strict
@@ -12,6 +13,7 @@
 
 ### Refactor
 
+- just run pre-commit install
 - replace pass with ... to keep it from being removed
 
 ## 2021.0.6.0 (2021-10-31)