+++ docs/docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Feat
 
+- add cct command
 - initialize the cement CLI application
 - replace toml with tomli (#74)
 - move wily to nox to reduce version conflicts
@@ -10,6 +11,7 @@
 
 ### Fix
 
+- cct was not being created as an alias. Use full command
 - wily
 
 ### Refactor