+++ docs/docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+## 2021.0.7.0 (2021-12-31)
+
 ### Feat
 
 - add cct command